Added context for hash get so we can timeout etc
